+++ tests/States/Loader/FinderComposerTest.php
@@ -29,7 +29,7 @@ use UniAlteri\States\Loader\Exception;
 use UniAlteri\Tests\Support;
 
 /**
- * Class FinderStandardTest
+ * Class FinderComposerTest
  * Tests the excepted behavior of standard finder implementing the interface \UniAlteri\States\Loader\FinderInterface.
  *
  * @copyright   Copyright (c) 2009-2015 Uni Alteri (http://agence.net.ua)
@@ -40,7 +40,7 @@ use UniAlteri\Tests\Support;
  * @license     http://teknoo.it/states/license/gpl-3.0     GPL v3 License
  * @author      Richard Déloge <r.deloge@uni-alteri.com>
  */
-class FinderStandardTest extends \PHPUnit_Framework_TestCase
+class FinderComposerTest extends \PHPUnit_Framework_TestCase
 {
     /**
      * Finder to test.
@@ -156,12 +156,12 @@ class FinderStandardTest extends \PHPUnit_Framework_TestCase
      * @param string $statedClassName
      * @param string $pathString
      *
-     * @return Loader\FinderStandard
+     * @return Loader\FinderComposer
      */
     protected function initializeFinder($statedClassName, $pathString)
     {
         $virtualDIContainer = new Support\MockDIContainer();
-        $this->finder = new Loader\FinderStandard($statedClassName, $pathString, $this->getClassLoaderMock());
+        $this->finder = new Loader\FinderComposer($statedClassName, $pathString, $this->getClassLoaderMock());
         $this->finder->setDIContainer($virtualDIContainer);
 
         return $this->finder;
@@ -172,7 +172,7 @@ class FinderStandardTest extends \PHPUnit_Framework_TestCase
      */
     public function testSetAndGetDiContainer()
     {
-        $object = new Loader\FinderStandard('', '', $this->getClassLoaderMock());
+        $object = new Loader\FinderComposer('', '', $this->getClassLoaderMock());
         $virtualContainer = new Support\MockDIContainer();
         $this->assertSame($object, $object->setDIContainer($virtualContainer));
         $this->assertSame($virtualContainer, $object->getDIContainer());